Frontend progress before final report

+++ frontend/src/components/SignupLogin.vue
@@ -7,7 +7,7 @@
             <img src="../assets/logo4.png" alt="title-image" style="width: 35%;">
         </template>
         <template v-if="loginVisible" v-slot:default>
-            <el-form ref="form" :model="loginForm" :rules="rules">
+            <el-form ref="loginForm" :model="loginForm" :rules="loginRules">
                 <el-form-item
                 label=""
                 prop="email"
@@ -30,34 +30,22 @@
             </span>
         </template>
         <template v-if="signupVisible" v-slot:default>
-            <el-form ref="form" :model="registerForm" :rules="rules">
+            <el-form ref="registerForm" :model="registerForm" :rules="registerRules">
                 <el-form-item
                     label=""
                     prop="email"
-                    :rules="[
-                    { required: true, message: '请输入邮箱', trigger: 'blur' },
-                    { type: 'email', message: '请输入正确的邮箱格式', trigger: ['blur', 'change'] }
-                    ]"
                 >
                     <el-input id="email" prefix-icon="el-icon-message" placeholder="Email" v-model="registerForm.email"></el-input>
                 </el-form-item>
                 <el-form-item
                     label=""
                     prop="password"
-                    :rules="[
-                    { required: true, message: '请输入密码', trigger: 'blur' },
-                    { min: 6, max: 20, message: '密码长度在 6 到 20 个字符', trigger: 'blur' }
-                    ]"
                 >
                     <el-input id="password" prefix-icon="el-icon-lock" placeholder="Password" type="password" v-model="registerForm.password"></el-input>
                 </el-form-item>
                 <el-form-item
                     label=""
                     prop="confirmPassword"
-                    :rules="[
-                    { required: true, message: '请确认密码', trigger: 'blur' },
-                    { validator: validateConfirmPassword, trigger: 'blur' }
-                    ]"
                 >
                     <el-input id="confirmPassword" prefix-icon="el-icon-s-check" placeholder="Confirm password" type="password" v-model="registerForm.confirmPassword"></el-input>
                 </el-form-item>
@@ -80,7 +68,7 @@ import axios from 'axios';
 // import VueCookies from 'vue-cookies';
 import { useStore } from 'vuex';
 import sha256 from 'crypto-js/sha256';
-import { Msgbox } from 'element3';
+import { Msgbox, Message } from 'element3';
 // import aes from "crypto-js/aes";
 const store = useStore()
 // var SHA256 = require("crypto-js/sha256");
@@ -111,18 +99,44 @@ const loginForm = ref({
     email: '',
     password: '',
 })
-const loginFormRef = ref(null);
+var validateConfirmPassword = (rule, value, callback) => {
+        if (value === '') {
+          callback(new Error('Please enter password again'))
+        } else if (value !== registerForm.password) {            
+          callback(new Error('The passwords entered twice are inconsistent!'))
+        } else {
+          callback()
+        }}
+// const loginFormRef = ref(null);
 const loginRules = ref({
-    email: [{ required: true, message: 'please input email', trigger: 'blur' }],
-    password: [{ required: true, message: '请输入密码', trigger: 'blur' }]
+    email: [{ required: true, message: 'Please input email', trigger: 'blur' },
+    { type: 'email', message: 'Please enter the correct email format', trigger: ['blur', 'change'] }],
+    password: [{ required: true, message: 'please input password', trigger: 'blur' },
+    { min: 4, max: 20, message: 'Password length between 4 and 20 characters', trigger: 'blur' }]
+});
+const registerRules = ref({
+    email: [{ required: true, message: 'Please input email', trigger: 'blur' },
+    { type: 'email', message: 'Please enter the correct email format', trigger: ['blur', 'change'] }],
+    password: [{ required: true, message: 'please input password', trigger: 'blur' },
+    { min: 4, max: 20, message: 'Password length between 4 and 20 characters', trigger: 'blur' }],
+    confirmPassword: [{ required: true, message: 'Please confirm your password', trigger: 'blur' },
+                    { validator: validateConfirmPassword, trigger: 'blur' }]
 });
 function loginSubmit() {
+    // loginForm.value.validate((valid)=>{
+    //     if (valid){
+    //         console.log('valid')
+
+    //     } else{
+    //         console.log('not valid')
+    //     }
+    // })
     axios.post(store.state.url+'/auth/login', {
         'email': loginForm.value.email,
         'password': sha256(loginForm.value.password).toString()
     })
     .then(res => {
-        console.log('login...')
+        // console.log('login...')
         localStorage.setItem('token', res.data.data.token)
         localStorage.setItem('userid', res.data.data.userid)
 
@@ -132,8 +146,16 @@ function loginSubmit() {
 
         emit('update:loginVisible', false);
         emit('update:signupVisible', false)
+        Message({
+            message: 'Login success!',
+            type: 'success'
+          })
     })
     .catch( error => {
+        Message({
+            message: 'Login failure...',
+            type: 'warning'
+          })
         console.error(error)
     })
 }
@@ -145,17 +167,36 @@ function signupSubmit() {
     // console.log(sha256('aabb').toString())
     // console.log(localStorage.getItem('token')==null)
     // console.log(aes.encrypt('my message', 'secret key 123').toString())
+    // registerForm.value.validateField('confirmPassword').then(res => {
+    //     console.log(res)
+    // }).catch(({errors, fields}) => {
+    //     console.log('error:', errors)
+    //     console.log('fields:', fields)
+    // })
+    // if (registerForm.confirmPassword !== registerForm.password) {
+    //     callback(new Error('The passwords entered twice are inconsistent!'))
+    //     // console.log
+    // }
+    // if (registerForm.value.validateField('confirmPassword')) {
+    //     console.log('yes')
+        
+    // } else{
+    //     console.log('no')
+    //     // registerForm.resetFields('confirmPassword')
+    //     console.log(a)
+
+    // }
     axios.post(store.state.url+'/auth/register', {
         "email": registerForm.value.email,
         "password": sha256(registerForm.value.password).toString()
     })
     .then(res => {
-        console.log('signup...')
-        console.log('login success:', res.data);
+        // console.log('signup...')
+        // console.log('login success:', res.data);
         // 登录成功，跳转到首页或其他页面
         localStorage.setItem('token', res.data.data.token)
         localStorage.setItem('userid', res.data.data.userid)
-        console.log(res.data.data.wallet)
+        // console.log(res.data.data.wallet)
 
         store.state.logState = true
         store.state.token = res.data.data.token
@@ -169,12 +210,20 @@ function signupSubmit() {
         emit('update:signupVisible', false);
         privateKeyAlert(res.data.data.wallet.privateKey)
         saveAsTxtFile(res.data.data.wallet.privateKey, 'privateKey.txt')
+        Message({
+            message: 'Signup success!',
+            type: 'success'
+          })
     })
     .catch( error => {
+        Message({
+            message: 'Signup failure...',
+            type: 'warning'
+          })
         console.error(error)
     })
     .then(()=>{
-        console.log('signup...')
+        // console.log('signup...')
         
     })
     // .catch(error => {
